Do not explicitly require rubygems
RubyGems should not be explicitly required by gems themselves. If the
gem is being used, rubygems (or another gem manager) is already at
work. Also it is reasonable for a library/gem to assume that LOAD_PATH
contains the usual lib/ directory.

This post explains the issue at stake more precisely:
<http://tomayko.com/writings/require-rubygems-antipattern>.