Consolidate all files searches in a function
In various places there is the need to search for some files and
subsequently filter out some of the found files. The new method
Enumerable#find_files_and_filter searches each directory of the
enumerable object, finds all the files that match a certain pattern
and filters out all for which one of supplied regexp filters match.