Generate a file task for each Ruby file
Instead of using an indirect task to compile the Ruby files, we generate
a file task for each Ruby file. With this change, Ruby files are
compiled only if changed since the last run or Rawr.

Because of the API exposed by JRuby 1.4.0, there is no way to compile
a single ruby file so all the classes are recompiled even if only one
Ruby file needs to be recompiled.